pyproject.toml
[project]
name = "boxctl"
version = "0.3.6"
description = "Secure, isolated Docker containers for AI coding agents with complete development tooling"
readme = "README.md"
authors = [{name = "Marc Schuetze", email = "marc@schuetze.io"}]
requires-python = ">=3.10,<4.0"
dependencies = [
    "click>=8.1.7,<9.0.0",
    "docker>=7.0.0,<8.0.0",
    "pyyaml>=6.0.3,<7.0.0",
    "rich>=13.7.0,<15.0.0",
    "tomli>=2.0.1,<3.0.0",
    "toml>=0.10.2,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "anthropic>=0.39.0,<1.0.0",
    "questionary>=2.1.1,<3.0.0",
    "asyncssh>=2.14.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
]

[project.scripts]
boxctl = "boxctl.cli:main"
box = "boxctl.cli:main"
abox = "boxctl.cli:main"
agentctl = "boxctl.agentctl:main"

[tool.poetry]
packages = [{include = "boxctl"}]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
black = "^24.3.0"
mypy = "^1.7.1"
ruff = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_network: marks tests that need network access",
    "requires_docker: marks tests that need Docker",
    "pulls_images: marks tests that pull Docker images",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
